Move logging config to api.py
